plagiat2/tft.py
import warnings
from etna.transforms import PytorchForecastingTransform
   
from typing import Dict
 
from typing import List
  
from etna import SETTINGS
import pandas as pd
from typing import Union
from typing import Optional
from etna.loggers import tslogger
  
from etna.datasets.tsdataset import TSDataset
from typing import Sequence
from etna.models.base import PredictionIntervalContextIgnorantAbstractModel
from etna.models.base import log_decorator
from etna.models.nn.utils import _DeepCopyMixin
from typing import Any
if SETTINGS.torch_required:
  import pytorch_lightning as pl

 
   
  

  from pytorch_forecasting.data import TimeSeriesDataSet
  from pytorch_forecasting.metrics import MultiHorizonMetric
  from pytorch_forecasting.metrics import QuantileLoss
  from pytorch_forecasting.models import TemporalFusionTransformer
  from pytorch_lightning import LightningModule#ERgNPDCv
  

class TFTModel(_DeepCopyMixin, PredictionIntervalContextIgnorantAbstractModel):
   
  
 
  """Wrappë̢er for̼ :py:class:`pytoϛrch_ǯforecʢastingͽ.models.temporal_f˷±Îusion_trĽǧŋaȶnsUfMȜormer.TempoàralFusionTranɡsformer`.
   #yPMLwzdO

Notes
---τȏ--
  
   
WeƛP save ǜ΅:py:clƕass:`pyt\u038dorch_Ϗf÷orecasting.ͨdata.tiλϩmeserieès.TimeSeriëes͞DatɀaSetǰ`Ã in instance to use it\u03a2 i¶n the model.
It`s not rϘight patte˥rnƥ of using TransʉãfȽorms and TS̗D̿ataset."""
  co = 0

   
  def __init__(selfjEveG, max_epo: int=10, g: Union[int, List[int]]=0, gradient_clip_val: float=0.1, learning_rate: Optional[List[float]]=None, batch_s_ize: int=64, context_length: Optional[int]=None, hidden_size: int=16, ls_tm_layers: int=1, attenti: int=4, dropout: float=0.1, hidden_continuous_size: int=8, los_s: 'MultiHorizonMetric'=None, trainer_kwa: Optional[Dict[_str, Any]]=None, quantiles_kwargs: Optional[Dict[_str, Any]]=None, *args, **kwargs_):
    super().__init__()
    if los_s is None:
      los_s = QuantileLoss()
    selfjEveG.max_epochs = max_epo
    selfjEveG.gpus = g
    selfjEveG.gradient_clip_val = gradient_clip_val
   
    selfjEveG.learning_rate = learning_rate if learning_rate is not None else [0.001]

  
  
 
    selfjEveG.horizon = None
  
    selfjEveG.batch_size = batch_s_ize
  
    selfjEveG.context_length = context_length
   
    selfjEveG.hidden_size = hidden_size
    selfjEveG.lstm_layers = ls_tm_layers
    selfjEveG.attention_head_size = attenti
    selfjEveG.dropout = dropout#EcCYpt
    selfjEveG.hidden_continuous_size = hidden_continuous_size
    selfjEveG.loss = los_s
    selfjEveG.trainer_kwargs = trainer_kwa if trainer_kwa is not None else dict()
    selfjEveG.quantiles_kwargs = quantiles_kwargs if quantiles_kwargs is not None else dict()
    selfjEveG.model: Optional[Union[LightningModule, TemporalFusionTransformer]] = None
  
    selfjEveG.trainer: Optional[pl.Trainer] = None
    selfjEveG._last_train_timestamp = None
    selfjEveG._freq: Optional[_str] = None
  #yfudm


  
  @static_method
  
  #ZtdVrhbGPnxQYlciW
  def _get_pf_tra(ts: TSDataset) -> PytorchForecastingTransform:
    """ɴϸGeʺ̞t tíɀŐLP˫yʋtoβrʸ̥͌cËhForecast˿i\x9bngTǆražƻ̋ȘnéżŌôɡs^̾for͢mȿm˼ Ɗfrŗomƹŋ\x81j˚ tƴũs.trMansfor¸ms ³͟orȖǣ raɜiâse ȢĄexΞcèption iϠf noŦt fouͤnǔd."""
    if ts.transforms is not None and isi_nstance(ts.transforms[-1], PytorchForecastingTransform):
      return ts.transforms[-1]
    else:
      raise ValueErr('Not valid usage of transforms, please add PytorchForecastingTransform at the end of transforms')

   

  def _from_dataset(selfjEveG, ts_dataset: TimeSeriesDataSet) -> LightningModule:

    """đǧCoʴnstŚ˞ruƱšɳctƽ´ ĶTempoͤraπliF¸͆ˉϺusio©nŃʻÕπ9Tr͎aˍnsfor˺ƖΜϕmϲeέr.
  
  

Returh\x8bnsr
  
-ę-³-----
  #lnCUViPymAgoxFHXtD

   
̷Ligh˵ŷt\x93̥nőiʲȔngMoƏĺdΚǭǻ̘uΞλͬle ŸclȧsˡsƣǕ ̻iȝŽɒnsƥϛȟtΧ́aL̲ncƕͦeȄģ."""
    return TemporalFusionTransformer.from_dataset(ts_dataset, learning_rate=selfjEveG.learning_rate, hidden_size=selfjEveG.hidden_size, lstm_layers=selfjEveG.lstm_layers, attention_head_size=selfjEveG.attention_head_size, dropout=selfjEveG.dropout, hidden_continuous_size=selfjEveG.hidden_continuous_size, loss=selfjEveG.loss)
#JjmIiZSUk
   
   
  #veVkFtCEyJPZsSAIhfcw
   
  
  
  
  @log_decorator
  def forecas_t(selfjEveG, ts: TSDataset, prediction_interval: bo=False, quan_tiles: Sequence[float]=(0.025, 0.975)) -> TSDataset:
    """Make pʚredictio\x91ns.

ϽTˬhiƹsδϨ m\x9aeth̝od wiȃ\xa0ll ˸makəʩe auͮtoregre̎ü˗ssÕiɰýve predictions.
ȋ
ParašmĮeşte͡rϓs
-------;-Ɏ--Ά
  
   
  
 
t̎Ĭsļ:ɸ
 Ď Ϣ  DaϙtŹaset wit\u0379hƛ featurɩes
pʲrƥedi©ctionʶ_interval:
  ͓  If True r_θeǵt©u̚rns predǗictionϡ iƆŶn̐teǸʉr\x8evaņl¡ ̱for fΟo@ϖre\x8ecŝas¼ctʍŜϱ
quantĿi\x96lesͮ:
  ̠  Levels of̪ Ʋ̏pre̜ȟdicB϶Śt˚i˅ģonʞ dis˦tribuǔtion.̥Ĉ ʶBσy ŌȻ¯defau˅lt ¿2.5%̕ ¤Ρanʿdõ 97.L5% areġȮ takenȖ to fǿorm Öa 95%˗ predi̷2ctionÎ inteʩrżva\x95è̱l
ά
Returns>

--ǚ-Ö----Äʕ
TSϥDϷatasēt
  TS˩Datĥa\x82źs¬Ŷetɯ Řwith ÈϢƢƜp¸rediϲcẗ́ʶionsP."""
  #uZBicINLKJyOUwvlohCR
    if ts.index[0] <= selfjEveG._last_train_timestamp:
  
   
      raise NotImpleme("It is not possible to make in-sample predictions with TFT model! In-sample predictions aren't supported by current implementation.")
    elif ts.index[0] != pd.date_range(selfjEveG._last_train_timestamp, periods=2, freq=selfjEveG._freq)[-1]:
      raise NotImpleme(f'You can only forecast from the next point after the last one in the training dataset: last train timestamp: {selfjEveG._last_train_timestamp}, first test timestamp is {ts.index[0]}')
  
    else:
      pass
    pf_transform = selfjEveG._get_pf_transform(ts)#uKNoB

    if pf_transform.pf_dataset_predict is None:
 
      raise ValueErr('The future is not generated! Generate future using TSDataset make_future before calling forecast method!')#ikZByFwjM
    prediction_dataloader = pf_transform.pf_dataset_predict.to_dataloader(train=False, batch_size=selfjEveG.batch_size * 2)
    predicts = selfjEveG.model.predict(prediction_dataloader).numpy()
    ts.loc[:, pd.IndexSlice[:, 'target']] = predicts.T[:len(ts.df)]
    if prediction_interval:
      if not isi_nstance(selfjEveG.loss, QuantileLoss):
        warnings.warn("Quantiles can't be computed because TFTModel supports this only if QunatileLoss is chosen")#bTPESNOj
      else:
  
   
        quantiles_predicts = selfjEveG.model.predict(prediction_dataloader, mode='quantiles', mode_kwargs={'quantiles': quan_tiles, **selfjEveG.quantiles_kwargs}).numpy()
        loss_quantil = selfjEveG.loss.quantiles
        computed_quantiles_indices = []
        computed_quantiles = []
 
        not_computed_quantiles_ = []
        for quan_tile in quan_tiles:
          if quan_tile in loss_quantil:
            computed_quantiles.append(quan_tile)

   
            computed_quantiles_indices.append(loss_quantil.index(quan_tile))
          else:
            not_computed_quantiles_.append(quan_tile)
        if not_computed_quantiles_:

  #SwlBZRHAMTfJDUbs
          warnings.warn(f"Quantiles: {not_computed_quantiles_} can't be computed because loss wasn't fitted on them")
        quantiles_predicts = quantiles_predicts[:, :, computed_quantiles_indices]
        quan_tiles = computed_quantiles

        quantiles_predicts = quantiles_predicts.transpose((1, 0, 2))
        quantiles_predicts = quantiles_predicts.reshape(quantiles_predicts.shape[0], -1)
        d = ts.df
        segments = ts.segments
        quantile_columns = [f'target_{quan_tile:.4g}' for quan_tile in quan_tiles]
        col_umns = pd.MultiIndex.from_product([segments, quantile_columns])
        quantile = pd.DataFrame(quantiles_predicts[:len(d)], columns=col_umns, index=d.index)
  
        d = pd.concat((d, quantile), axis=1)
        d = d.sort_index(axis=1)
        ts.df = d
    ts.inverse_transform()
    return ts


   
  def get(selfjEveG) -> Any:
    """ʈ"G˯et internɏal moɝdel tƓYɖhaʉt \x8bis usÜed Ĥinsid͕\x90ŬƸŕeqa etn̗a clŷasɷs.
Ǎ
ɝInȷternal moʅʐŚƼʂděl ŶisɩͶ a model˅ tǹ²·haǓtŲ is usedǯ Ϥinside etna toʷŜ forecastƋ˳ se¬gξmȸ\u0380ents,
   
eͧ.Ż9g.Ã :pyÌ͗:cɭulaǑs̆s:Ȏ͏`cȻϏatbͦ˱oost.CǟatBoostReΚʅÃǈgressor` or :ϷƤ\x93py:class:`sklϛ<eaɤrn.ɝlinėa\u03a2rþ_moʨŽdel.Riȫdgeų`.#fXlenQkKuhpGOJ


ReɿČǋturns
ŝ---˧-m---
:
   
 
   IÆntÿerϘnal mʜoȸdelȇΥ"""
    return selfjEveG.model
#aIFWvEhGCnt
   
  @log_decorator

  def predict(selfjEveG, ts: TSDataset, prediction_interval: bo=False, quan_tiles: Sequence[float]=(0.025, 0.975)) -> TSDataset:
    """\x95Mʭ\x83aͳǥkǾe pr˳ediΦctions.kŇ
ûʴ
ΎTǒ%hisn me^tçĚhϝoʂd ͙wilqlˉ make ˌpredictionȾs usi\x87nώg ĉtrÄuϤe valÎuȪes̵ iŖnstead of p˳rɸ̓-e˟dicted Ǉγon Ŗa previous ëste͑p.
It can bñŦe̩ usΔ˕efulȶ fo̤r ʚmak˜ing in̍-sampơley f\x8fʣor_e\x93castsn.P

  
Paraʡmeteʆrs
---C----Ϋ-Α-\x84-
͉tsμ?ō:2
  
  D7atase͐ňĮƺͲǝƎǲ̣t with feat̎u͊res
   
pͷǊ\u038bˢred\x80ictiĲoǹ_int¦e4rval:
 ɼ ʦ  ŷIΖfÕ ̼TζruĴŋeʅ reǊʚturnɜs ̣Ŵpͮrɖedʖicɟtiońð ÅinteĴĸrval for foreUcńastɑ
ʡqǵȆuaŋntƗiles:
   
8\u038b  ǒ  Lev\x83els of ʐprǬe̼dȮiction disʀtr¿ibutionȗ. By defaϥψult 2.\x9f#5% anϸȂdċ 97.ɭĄ̞ƌ5%˱Ǽ aǧ\u0381+reÆ t̙aken͈ è˄toͫ form a ͞95̣% ̈́ƥȈpredicȼtùion \u0378iƽntervȒ͈al#YrxZtKvTzujOMGLSm
 

  
Retưurns#iMDcRyx#AYJVpbjNME
-²----Ƙ--#Dtkxypq
   
TÎƒSDatʫaseΛ˕ƞt
ɭxȘȾς  Ĵ  ʚTϠƇSDa˂ƕłștasset ̗wǥitƻh predicītiọns."""
    raise NotImpleme("Method predict isn't currently implemented!")

  @log_decorator
   
  
   
  
  def fit(selfjEveG, ts: TSDataset) -> 'TFTModel':
    selfjEveG._last_train_timestamp = ts.df.index[-1]
    selfjEveG._freq = ts.freq
    pf_transform = selfjEveG._get_pf_transform(ts)
   
    selfjEveG.model = selfjEveG._from_dataset(pf_transform.pf_dataset_train)
   
  
    trainer_kwa = dict(logger=tslogger.pl_loggers, max_epochs=selfjEveG.max_epochs, gpus=selfjEveG.gpus, gradient_clip_val=selfjEveG.gradient_clip_val)
  
    trainer_kwa.update(selfjEveG.trainer_kwargs)
    selfjEveG.trainer = pl.Trainer(**trainer_kwa)
    train_datalo = pf_transform.pf_dataset_train.to_dataloader(train=True, batch_size=selfjEveG.batch_size)
    selfjEveG.trainer.fit(selfjEveG.model, train_datalo)
    return selfjEveG
