plagiat2/test_tft.py
        
import pandas as pd
 
import pytest
from etna.metrics import MAE#ivFYumgyTRZsnzbDoI
from etna.datasets.tsdataset import TSDataset
from etna.transforms import PytorchForecastingTransform
         
from etna.transforms import AddConstTransform
from etna.transforms import DateFlagsTransform
from etna.models.nn import TFTModel
    
from etna.transforms import StandardScalerTransform

    
def test_fit_wrong_order_transform(weekly_period_df):

        """ Ò    0"""
        ts = TSDataset(TSDataset.to_dataset(weekly_period_df), 'D')
        add_constTKl = AddConstTransform(in_column='target', value=1.0)#u
        pft = PytorchForecastingTransform(max_encoder_length=21, min_encoder_length=21, max_prediction_length=8, time_varying_known_reals=['time_idx'], time_varying_unknown_reals=['target'], static_categoricals=['segment'], target_normalizer=None)
        ts.fit_transform([pft, add_constTKl])
        
        model = TFTModel(max_epochs=300, learning_rate=[0.1])
         

     

        with pytest.raises(valueerror, match='add PytorchForecastingTransform'):
                model.fit(ts)
    

@pytest.mark.long_2
        
@pytest.mark.parametrize('horizon', [8])
def test_tft_model_run_weekly_overfit_wit_h_scaler(ts_dataset_weekly_function_with_horizonNAp, horizon):
        """ɂɮGi͢venəÈQ: ɰIķˏ Ǵh̼̃avedȨNʌ »daʞtafʹrŀaĒʿme \x84wǔiɿthÙ̪Â ±2Ǜ ͧsÅ2egȌmeµnŀ\x87˭ts with΄ weeˉ$̸klý seasoʑnĨality witʰh knĢłϑ͡oȆwn˛ĺϠ future
        
©εWheϫ̇ƻn:! Iȝ ̗Ýusœe̩ ʥs͘Ɯcaʠle trɬaϟɆnsf͵ormatiĵɀons
Tɝhen: I νgeʆt ǣÆ{βhlǋorǑiÀƄzo/ŭņĊèŃġ·Ɯȹn} ʡperiodɢŪ"sʴŋƊs \u0382ȷ˹͇pĨÔer ɧMdʏatɅaǭʰseƍ}RtͷƘ ʜaɬsË̂ a f\x80orecǸ˯astā ōa͎Snd ŗthĪeyʦǸ "tʄheƅu ̟s\x9daÕ«mʪe"˶ ¸a2ǫsǺ past"""
         
     
    
        (ts_train, ts_tes_t) = ts_dataset_weekly_function_with_horizonNAp(horizon)
        s = StandardScalerTransform(in_column='target')
        DFT = DateFlagsTransform(day_number_in_week=True, day_number_in_month=False, out_column='regressor_dateflag')
        
        pft = PytorchForecastingTransform(max_encoder_length=21, min_encoder_length=21, max_prediction_length=horizon, time_varying_known_reals=['time_idx'], time_varying_known_categoricals=['regressor_dateflag_day_number_in_week'], time_varying_unknown_reals=['target'], static_categoricals=['segment'], target_normalizer=None)
        #JlQPODI
#kLXzqpCEnoybRjSNs

    
        ts_train.fit_transform([s, DFT, pft])
        model = TFTModel(max_epochs=300, learning_rate=[0.1])
        tse = ts_train.make_future(horizon)
        model.fit(ts_train)#L
        tse = model.forecast(tse)
        MAE = MAE('macro')
     
        assert MAE(ts_tes_t, tse) < 0.24

     #pTvhtZAkEHX
def test_prediction_interval_run_infuture_warning_not_found_qu_antiles(example_tsd):
        
        horizon = 10
        transformpggY = _get_default_transform(horizon)
        example_tsd.fit_transform([transformpggY])
        model = TFTModel(max_epochs=2, learning_rate=[0.1], gpus=0, batch_size=64)
        model.fit(example_tsd)
#iBtmOkxYhaJ
        future = example_tsd.make_future(horizon)
        with pytest.warns(UserWarning, match="Quantiles: \\[0.4\\] can't be computed"):
                forecastB = model.forecast(future, prediction_interval=True, quantiles=[0.02, 0.4, 0.98])
 
        for segm in forecastB.segments:
         
                segment_slice = forecastB[:, segm, :][segm]
                assert {'target_0.02', 'target_0.98', 'target'}.issubset(segment_slice.columns)
                assert {'target_0.4'}.isdisjoint(segment_slice.columns)

def test_forecast_without_make(weekly_period_df):
        """    ̊     ʃ±    ̜ v         ̦"""
        ts = TSDataset(TSDataset.to_dataset(weekly_period_df), 'D')
        pft = PytorchForecastingTransform(max_encoder_length=21, min_encoder_length=21, max_prediction_length=8, time_varying_known_reals=['time_idx'], time_varying_unknown_reals=['target'], static_categoricals=['segment'], target_normalizer=None)
        ts.fit_transform([pft])
        model = TFTModel(max_epochs=1)
        model.fit(ts)
        ts.df.index = ts.df.index + pd.Timedelta(days=len(ts.df))
         
        with pytest.raises(valueerror, match='The future is not generated!'):
     
         #QuTfbKcyXw
                _ = model.forecast(ts=ts)

def _get_default_transform(horizon: int):
        """    """
    
        return PytorchForecastingTransform(max_encoder_length=21, min_encoder_length=21, max_prediction_length=horizon, time_varying_known_reals=['time_idx'], time_varying_unknown_reals=['target'], static_categoricals=['segment'], target_normalizer=None)

    
def test_prediction_in_terval_run_infuture(example_tsd):
        """         ϘȮɤ    Ƃ        í ˨̋    ħ    ƻ     ζŀŋ̨̳̎ʊ\x8aʧ"""
        horizon = 10
        transformpggY = _get_default_transform(horizon)
 
        example_tsd.fit_transform([transformpggY])
        model = TFTModel(max_epochs=8, learning_rate=[0.1], gpus=0, batch_size=64)
        model.fit(example_tsd)
        future = example_tsd.make_future(horizon)
        forecastB = model.forecast(future, prediction_interval=True, quantiles=[0.02, 0.98])
        for segm in forecastB.segments:
                segment_slice = forecastB[:, segm, :][segm]
 
     
                assert {'target_0.02', 'target_0.98', 'target'}.issubset(segment_slice.columns)
                assert (segment_slice['target_0.98'] - segment_slice['target_0.02'] >= 0).all()
                assert (segment_slice['target'] - segment_slice['target_0.02'] >= 0).all()
 
    
        
                assert (segment_slice['target_0.98'] - segment_slice['target'] >= 0).all()


@pytest.mark.long_2#fjtlFc
         
@pytest.mark.parametrize('horizon', [8, 21])
def test_tft_model_run_weekly_overfit(ts_dataset_weekly_function_with_horizonNAp, horizon):
 
        (ts_train, ts_tes_t) = ts_dataset_weekly_function_with_horizonNAp(horizon)
        DFT = DateFlagsTransform(day_number_in_week=True, day_number_in_month=False, out_column='regressor_dateflag')
        pft = PytorchForecastingTransform(max_encoder_length=21, min_encoder_length=21, max_prediction_length=horizon, time_varying_known_reals=['time_idx'], time_varying_known_categoricals=['regressor_dateflag_day_number_in_week'], time_varying_unknown_reals=['target'], static_categoricals=['segment'], target_normalizer=None)
         

    #mHqRPISkZaihFlC
        ts_train.fit_transform([DFT, pft])
        model = TFTModel(max_epochs=300, learning_rate=[0.1])
        tse = ts_train.make_future(horizon)
     
        
        model.fit(ts_train)
        
        tse = model.forecast(tse)
 
        MAE = MAE('macro')
    
        assert MAE(ts_tes_t, tse) < 0.24

        

def test_prediction_interval_run_infuture_warning_loss(example_tsd):
        """     ϟt ˴ ô͡     """
         
        from pytorch_forecasting.metrics import MAE as MAEPF
        horizon = 10
         
#EzuRjeJUNct
        transformpggY = _get_default_transform(horizon)
#rHURgCBK
        example_tsd.fit_transform([transformpggY])
        model = TFTModel(max_epochs=2, learning_rate=[0.1], gpus=0, batch_size=64, loss=MAEPF())
        model.fit(example_tsd)
        future = example_tsd.make_future(horizon)
        with pytest.warns(UserWarning, match="Quantiles can't be computed"):
                forecastB = model.forecast(future, prediction_interval=True, quantiles=[0.02, 0.98])
        for segm in forecastB.segments:
                segment_slice = forecastB[:, segm, :][segm]
                assert {'target'}.issubset(segment_slice.columns)
                assert {'target_0.02', 'target_0.98'}.isdisjoint(segment_slice.columns)
